+++ src/Nethermind/Nethermind.Core.Test/BlockHeaderTests.cs
@@ -40,7 +40,7 @@ namespace Nethermind.Core.Test
             header.GasUsed = (long)Bytes.FromHexString("0x5208").ToUnsignedBigInteger();
             header.MixHash = new Keccak(Bytes.FromHexString("0x00be1f287e0911ea2f070b3650a1a0346535895b6c919d7e992a0c255a83fc8b"));
             header.Nonce = (ulong)Bytes.FromHexString("0xa0ddc06c6d7b9f48").ToUnsignedBigInteger();
-            header.Number = Bytes.FromHexString("0x01").ToUInt256();
+            header.Number = (long)Bytes.FromHexString("0x01").ToUInt256();
             header.ParentHash = new Keccak(Bytes.FromHexString("0x5a39ed1020c04d4d84539975b893a4e7c53eab6c2965db8bc3468093a31bc5ae"));
             header.ReceiptsRoot = new Keccak(Bytes.FromHexString("0x056b23fbba480696b65fe5a59b8f2148a1299103c4f57df839233af2cf4ca2d2"));
             header.StateRoot = new Keccak(Bytes.FromHexString("0x5c2e5a51a79da58791cdfe572bcfa3dfe9c860bf7fad7d9738a1aace56ef9332"));
@@ -65,7 +65,7 @@ namespace Nethermind.Core.Test
             header.GasUsed = (long)Bytes.FromHexString("0x5208").ToUnsignedBigInteger();
             header.MixHash = new Keccak(Bytes.FromHexString("0x615bbf44eb133eab3cb24d5766ae9617d9e45ee00e7a5667db30672b47d22149"));
             header.Nonce = (ulong)Bytes.FromHexString("0x4c4f3d3e055cb264").ToUnsignedBigInteger();
-            header.Number = Bytes.FromHexString("0x03").ToUInt256();
+            header.Number = (long)Bytes.FromHexString("0x03").ToUInt256();
             header.ParentHash = new Keccak(Bytes.FromHexString("0xde1457da701ef916533750d46c124e9ae50b974410bd590fbcf4c935a4d19465"));
             header.ReceiptsRoot = new Keccak(Bytes.FromHexString("0x056b23fbba480696b65fe5a59b8f2148a1299103c4f57df839233af2cf4ca2d2"));
             header.StateRoot = new Keccak(Bytes.FromHexString("0xfb4084a7f8b57e370fefe24a3da3aaea6c4dd8b6f6251916c32440336035160b"));
